[TTS] refactor nemo.collections.tts.data
