+++ nemo/collections/tts/g2p/models/ctc_g2p.py
@@ -24,7 +24,7 @@ from pytorch_lightning import Trainer
 from torch import nn
 from transformers import AutoConfig, AutoModel, AutoTokenizer
 
-from nemo.collections.tts.g2p.data.ctc_g2p import CTCG2PBPEDataset
+from nemo.collections.tts.data.g2p_dataset import CTCG2PBPEDataset
 from nemo.collections.tts.g2p.models.g2p_model import G2PModel
 from nemo.core.classes.common import PretrainedModelInfo
 from nemo.utils import logging
